Allow admins to associate events with panel applications more seamlessly (#124)
* Add a panelist ribbon to approved panelists without a guest badge
While testing other code, I noticed that if you made an existing group a 'guest' group and had them submit a panel, their POC would be a panelist but they were assumed to have a Guest badge. Just in case in the future we use the guest.html form to allow non-Guest-badge-owning panelists to submit new apps, I figured we should make sure they either have a guest badge or a panelist ribbon.

* Allow admins to associate events with panel applications more seamlessly
Right now, admins must create an event, then go to a totally different page to associate a panel application with that event. This seems silly, because it forces the admin to work in two separate places and to manually copy over panel names and descriptions. Now admins can associate events with panel applications while editing an event; additionally, if they associate a new event with a panel app, they can skip entering in name/desc. Any panelists associated with the panel app also get added to the event.

* Remove leftover commented-out line